Ensure EXP claim is stored as a Long in JWT token
Fixes gh-237

+++ spring-security-oauth2/src/main/java/org/springframework/security/oauth2/provider/token/DefaultAccessTokenConverter.java
@@ -82,7 +82,7 @@ public class DefaultAccessTokenConverter implements AccessTokenConverter {
 		info.remove(CLIENT_ID);
 		info.remove(SCOPE);
 		if (map.containsKey(EXP)) {
-			token.setExpiration(new Date((Integer) map.get(EXP) * 1000L));
+			token.setExpiration(new Date((Long) map.get(EXP) * 1000L));
 		}
 		if (map.containsKey(JTI)) {
 			info.put(JTI, map.get(JTI));
